--- src/live_stream_analysis/readadara/__init__.py
@@ -1,11 +1,0 @@
-# TODO: use readadara as depedency when own published packages instead of copy-paste; remove this module when ready
-
-from .adara_reader import AdaraFileReader, AdaraMultiFileReader, AdaraRunReader, AdaraLiveStreamReader, DoubleArchiver
-
-__all__ = [
-    "AdaraFileReader",
-    "AdaraMultiFileReader",
-    "AdaraRunReader",
-    "AdaraLiveStreamReader",
-    "DoubleArchiver",
-]
