--- tests/readadara/test_adara_reader.py
@@ -1,160 +1,0 @@
-"""Unit tests for live_stream_analysis.readadara."""
-
-from __future__ import annotations
-
-import io
-import struct
-import tempfile
-from pathlib import Path
-
-import pytest
-
-from live_stream_analysis.readadara.adara_reader import (
-    AdaraBankedEventPacket,
-    AdaraEventPacket,
-    AdaraFileReader,
-    AdaraRTDLPacket,
-    AdaraRawPacket,
-    read_all_packets,
-    read_packets_generator,
-)
-
-from .adara_fixtures import event_packet, null_packet, rtdl_packet
-
-
-# ---------------------------------------------------------------------------
-# Helpers
-# ---------------------------------------------------------------------------
-
-def _write_tmp(data: bytes, tmp_path: Path) -> Path:
-    p = tmp_path / "test.adara"
-    p.write_bytes(data)
-    return p
-
-
-# ---------------------------------------------------------------------------
-# Packet header parsing
-# ---------------------------------------------------------------------------
-
-class TestAdaraRawPacket:
-    def test_get_payload_length(self):
-        pkt = null_packet(format_int=0, timestamp_s=42)
-        p = AdaraRawPacket(pkt)
-        assert p.get_payload_length() == 0
-
-    def test_get_format_int(self):
-        pkt = null_packet(format_int=0x100, timestamp_s=0)
-        p = AdaraRawPacket(pkt)
-        assert p.get_format_int() == 0x100
-
-    def test_get_timestamp(self):
-        pkt = null_packet(format_int=0, timestamp_s=10)
-        p = AdaraRawPacket(pkt)
-        assert p.get_timestamp_s() == 10
-        assert p.get_timestamp() == pytest.approx(10.0)
-
-    def test_get_length_equals_header_plus_payload(self):
-        data = rtdl_packet(timestamp_s=5)
-        p = AdaraRawPacket(data)
-        # payload_length field + 16-byte header
-        assert p.get_length() == p.get_payload_length() + 16
-
-
-# ---------------------------------------------------------------------------
-# Event packet
-# ---------------------------------------------------------------------------
-
-class TestAdaraEventPacket:
-    def test_get_events_returns_list_of_tuples(self):
-        events_in = [(10, 200), (20, 500)]
-        data = event_packet(events_in, timestamp_s=3)
-        p = AdaraEventPacket(data)
-        events_out = p.get_events()
-        assert len(events_out) == len(events_in)
-        for (pid_in, tof_in), (pid_out, tof_out) in zip(events_in, events_out):
-            assert pid_in == pid_out
-            assert tof_in == tof_out
-
-    def test_empty_event_packet(self):
-        data = event_packet([], timestamp_s=1)
-        p = AdaraEventPacket(data)
-        assert p.get_events() == []
-
-
-# ---------------------------------------------------------------------------
-# RTDL packet
-# ---------------------------------------------------------------------------
-
-class TestAdaraRTDLPacket:
-    def test_rtdl_format_int(self):
-        data = rtdl_packet(timestamp_s=7, charge_10pc=50)
-        p = AdaraRTDLPacket(data)
-        assert p.get_format_int() == 0x100
-
-    def test_rtdl_charge(self):
-        # charge_10pc = 50 → get_charge() should be 50*10 = 500
-        data = rtdl_packet(timestamp_s=0, charge_10pc=50)
-        # AdaraRTDLPacket inherits get_charge from AdaraRawPacket which reads
-        # bytes [16+8:16+11] — not the RTDL-specific field.  We just assert the
-        # packet parses without error and returns an integer.
-        p = AdaraRTDLPacket(data)
-        assert isinstance(p.get_charge(), int)
-
-
-# ---------------------------------------------------------------------------
-# read_packets_generator / read_all_packets
-# ---------------------------------------------------------------------------
-
-class TestReadPackets:
-    def _make_stream(self, packets_bytes: list[bytes]) -> io.BytesIO:
-        combined = b"".join(packets_bytes)
-        buf = io.BytesIO(combined)
-        # BytesIO has no .name; provide a fake one so lazy path is skipped
-        return buf
-
-    def test_read_all_packets_count(self, tmp_path: Path):
-        data = rtdl_packet() + event_packet([(1, 1)]) + null_packet()
-        path = _write_tmp(data, tmp_path)
-        with path.open("rb") as f:
-            pkts = read_all_packets(f)
-        assert len(pkts) == 3
-
-    def test_read_packets_types(self, tmp_path: Path):
-        data = rtdl_packet() + event_packet([(5, 10)])
-        path = _write_tmp(data, tmp_path)
-        with path.open("rb") as f:
-            pkts = read_all_packets(f)
-        assert pkts[0].get_format_int() == 0x100
-        assert isinstance(pkts[1], AdaraEventPacket)
-
-    def test_empty_file_returns_empty(self, tmp_path: Path):
-        path = _write_tmp(b"", tmp_path)
-        with path.open("rb") as f:
-            pkts = read_all_packets(f)
-        assert pkts == []
-
-
-# ---------------------------------------------------------------------------
-# AdaraFileReader
-# ---------------------------------------------------------------------------
-
-class TestAdaraFileReader:
-    def test_read_populates_packets(self, tmp_path: Path):
-        data = rtdl_packet() + event_packet([(7, 77)])
-        path = _write_tmp(data, tmp_path)
-        reader = AdaraFileReader(str(path))
-        reader.read()
-        assert len(reader.packets) == 2
-
-    def test_read_generator_yields_packets(self, tmp_path: Path):
-        data = null_packet() + rtdl_packet() + null_packet()
-        path = _write_tmp(data, tmp_path)
-        reader = AdaraFileReader(str(path))
-        pkts = list(reader.read_generator())
-        assert len(pkts) == 3
-
-    def test_get_size(self, tmp_path: Path):
-        data = rtdl_packet()
-        path = _write_tmp(data, tmp_path)
-        reader = AdaraFileReader(str(path))
-        assert reader.get_size() == len(data)
